Fixed the following:
- If you write 'lift create ' it should list the templates instead of throwing an exception
- if it can't find the template it should list the templates instead of saying it can't find it
- Whenever it doesn't understand a command it should invoke help instead of saying it's an invalid command

And I'm almost done with:

- Switch to SBT's logging mechanism so we can get colors!

So far I've just removed all [error] etc in the messages and I've created a layer of abstraction for loggers so I can still run the commands in the console.
I'm currently not using the SBT logger because it threw a bunch of runtime exceptions when I did so I'll wait till I have the releases/snapshot set up I don't break everything for everyone.